backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cnn-vae-project"
version = "0.4.0"
description = "Convolutional Variational Autoencoder built with TensorFlow/Keras"
requires-python = ">=3.10"
dependencies = [
  "tensorflow>=2.15,<3.0",
  "numpy>=1.24,<3.0",
  "matplotlib>=3.8,<4.0",
  "datasets>=2.18,<4.0",
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pillow>=10.4,<12.0",
]

[tool.setuptools.packages.find]
include = ["backend", "backend.*"]
